Change insert_element() to return the type of the insertion call, rather than void     through std::void_t, since SFINAE std::void_t is a template alias and there are open     language issues on how template aliases should affect SFINAE.
    Make the insert_after() call use an interator passed by const reference instead of
    pointer so that it can be defaulted.

    Make the serialize_insertable.h code as common as possible, separating forward_list
    code out with "if constexpr".

    Generalize serialization to work with all types with certain traits

    Remove mapping name from serialization parameter list
    Add pushMapName() and popMapName() to get and set map name

+++ src/sst/core/serialization/impl/unpacker.h
@@ -19,6 +19,9 @@
 
 #include "sst/core/serialization/impl/ser_buffer_accessor.h"
 
+#include <cstddef>
+#include <string>
+
 namespace SST::Core::Serialization::pvt {
 
 class ser_unpacker : public ser_buffer_accessor
@@ -36,7 +39,7 @@ public:
      * @param buf   Must unpack to non-null buffer
      * @param size  Must be non-zero
      */
-    void unpack_buffer(void* buf, int size);
+    void unpack_buffer(void* buf, size_t size);
 
     void unpack_string(std::string& str);
 };
